refactor(transport): share SSRF policy resolver
